Final merge from dev into main (#12)
* bring Kamil's main changes into dev (#7)

* Dev (#6)

* Fixes for new domain, more frequent location updates

* Fix location trying to update when not signed in

* Added open search box, added max distance filter

* Fixes based on feedback (min age, other validations)

* Started discovery queue

* Make sideform sticky

* fix registration max-width

* Shrink header

* start chat styling

* chat list styling

* Personality Type input now only accepts letters, user age check to make sure you have to be 18 or over to sign up!

---------


* Added password requirements to registration page

---------


* Discovery queue and personalised suggestions

* Added unread and latest message to chat

* Added pagination to search

* Added ability to report profile

* fix chat merge

* Added admin dashboard

* Removed ide helper file, added admin middleware to admin dashboard

* Fixed chat styles again

* Delete Photo Functionality Added

* Styled the report button

* Fixed up the viewprofile page

* Admin only sees the ban form when going on someone sprofile

* DOin bits

* This is very good commit.


* Last fix :)


* Added trait selection functionality.

* typo fix for chat page style

* Added AI responses

* Admin updates, fixed ability to delete own account

* Chat updates, minor changes

* selected css applied to traits and interests buttons

* puuuuhs

* Added ability for admin to delete or edit user account, added unmatch button

* - Interests and traits finalized
- delete photo functionaity completed
- photo name field updated

* lil fix

* fix responsive sizes


* Changing style for Myers Briggs Link

* Improved admin dashboard

* Added notifications

* Changed discovery to take account of gender & interested in, removed self from search, changes to dashboard, chat is match only

* Fixed the navbar appearing at the top of viewprofile

* fixed the broken viewprofile page

* Styled the chat page

* Adding emoji to traits and interests in viewprofile

* Fixes, realtime notifications

* Fixes

* Photo link

* fixin

* More fixes (age 100+)

* Fixed interests + traits button

* Fixes

* Changes before code submission

---------

+++ app/Http/Controllers/NotificationController.php
@@ -12,8 +12,12 @@ class NotificationController extends Controller
     {
         $notif = Notification::find($req->id);
 
-        if ($notif->recipient_id == Auth::user()->id) {
-            $notif->delete();
+        if ($notif) {
+            if ($notif->recipient_id == Auth::user()->id) {
+                $notif->delete();
+                return response(null, 200);
+            }
         }
+        return response(null, 204);
     }
 }
